Change the debugger interactive window to use the debugger protocol.
