+++ Python/Product/PythonTools/ptvsd/attach_server.py
@@ -77,13 +77,6 @@ import ptvsd.visualstudio_py_ipcjson as vsipc
 #   From there on the socket is assumed to be using the normal PTVS debugging protocol.
 #   If attaching was not successful (which can happen if some other debugger is
 #   already attached), the server responds with accepted=False and closes the connection. 
-#
-# 'legacyRemoteReplAttach'
-#   Attach REPL to the process. If successful, the server responds with
-#   accepted=True, and from there on the socket is assumed to be using the
-#   normal PTVS REPL (binary) protocol. If not successful (which can happen if
-#   there is no debugger attached), the server responds with accepted=False and
-#   closes the connection. 
 
 DEFAULT_PORT = 5678
 PTVSDBG_VER = 7 # must be kept in sync with DebuggerProtocolVersion in PythonRemoteProcess.cs
@@ -367,35 +360,6 @@ class AttachLoop(vsipc.SocketIO, vsipc.IpcChannel):
                 vspd.mark_all_threads_for_break(vspd.STEPPING_ATTACH_BREAK)
 
                 _attached.set()
-
-                # Prevent from closing the socket, it will be used by debugger
-                self.__owned_socket = None
-            else:
-                self.send_response(
-                    request,
-                    accepted=False,
-                )
-        finally:
-            self.set_exit()
-
-    def on_legacyRemoteReplAttach(self, request, args):
-        if self.__waiting_for_authentication:
-            self.send_response(
-                request,
-                success=False,
-                message='legacyRemoteDebuggerAuthenticate request must be sent first.',
-            )
-            self.set_exit()
-            return
-
-        try:
-            if not vspd.DETACHED:
-                self.send_response(
-                    request,
-                    accepted=True,
-                )
-
-                vspd.connect_repl_using_socket(self.__owned_socket)
 
                 # Prevent from closing the socket, it will be used by debugger
                 self.__owned_socket = None
