+++ Python/Product/PythonTools/visualstudio_py_repl.py
@@ -350,13 +350,6 @@ actual inspection and introspection."""
         args = read_string(self.conn)
         self.execute_file_ex(filetype, filename, args)
 
-    def _cmd_debug_attach(self):
-        import visualstudio_py_debugger
-        port = read_int(self.conn)
-        id = read_string(self.conn)
-        debug_options = visualstudio_py_debugger.parse_debug_options(read_string(self.conn))
-        self.attach_process(port, id, debug_options)
-
     _COMMANDS = {
         to_bytes('run '): _cmd_run,
         to_bytes('abrt'): _cmd_abrt,
@@ -369,7 +362,6 @@ actual inspection and introspection."""
         to_bytes('inpl'): _cmd_inpl,
         to_bytes('excf'): _cmd_excf,
         to_bytes('excx'): _cmd_excx,
-        to_bytes('dbga'): _cmd_debug_attach,
     }
 
     def _write_member_dict(self, mem_dict):
@@ -377,10 +369,6 @@ actual inspection and introspection."""
         for name, type_name in mem_dict.items():
             write_string(self.conn, name)
             write_string(self.conn, type_name)
-
-    def on_debugger_detach(self):
-        with self.send_lock:
-            write_bytes(self.conn, ReplBackend._DETC)
 
     def init_debugger(self):
         from os import path
@@ -508,10 +496,6 @@ actual inspection and introspection."""
 
     def flush(self):
         """flushes the stdout/stderr buffers"""
-        raise NotImplementedError
-
-    def attach_process(self, port, debugger_id, debug_options):
-        """starts processing execution requests"""
         raise NotImplementedError
 
 def exit_work_item():
@@ -646,7 +630,7 @@ due to the exec, so we do it here"""
         # Otherwise we have a nested exception hanging around and the 2nd abort doesn't
         # work (that's probably an IronPython bug)
         try:    
-            new_modules = self._get_cur_module_set()
+            new_modules = get_cur_module_set()
             try:
                 if new_modules != cur_modules:
                     self.send_modules_changed()
@@ -707,11 +691,9 @@ due to the exec, so we do it here"""
                     sys.stderr.write('KeyboardInterrupt')
                 else:
                     # let IronPython format the exception so users can do -X:ExceptionDetail or -X:ShowClrExceptions
-                    exc_next = self.skip_internal_frames(exc_tb)
-                    sys.stderr.write(''.join(traceback.format_exception(exc_type, exc_value, exc_next)))
+                    print_exception_frames(exc_type, exc_value, exc_tb)
             else:
-                exc_next = self.skip_internal_frames(exc_tb)
-                sys.stderr.write(''.join(traceback.format_exception(exc_type, exc_value, exc_next)))
+                print_exception_frames(exc_type, exc_value, exc_tb)
 
             try:
                 self.send_error()
@@ -720,19 +702,6 @@ due to the exec, so we do it here"""
                 return True, None, None, None
 
         return False, cur_modules, cur_ps1, cur_ps2
-
-    def skip_internal_frames(self, tb):
-        """return the first frame outside of the repl/debugger code"""
-        while tb is not None and self.is_internal_frame(tb):
-            tb = tb.tb_next
-        return tb
-
-    def is_internal_frame(self, tb):
-        """return true if the frame is from internal code (repl or debugger)"""
-        f = tb.tb_frame
-        co = f.f_code
-        filename = co.co_filename
-        return filename.endswith('visualstudio_py_repl.py') or filename.endswith('visualstudio_py_debugger.py')
 
     def execution_loop(self):
         """loop on the main thread which is responsible for executing code"""
@@ -794,23 +763,6 @@ due to the exec, so we do it here"""
                 print(out_codec.decode(line, 'replace')[0].rstrip('\r\n'))
         except Exception:
             traceback.print_exc()
-
-    @staticmethod
-    def _get_cur_module_set():
-        """gets the set of modules avoiding exceptions if someone puts something
-        weird in there"""
-
-        try:
-            return set(sys.modules)
-        except:
-            res = set()
-            for name in sys.modules:
-                try:
-                    res.add(name)
-                except:
-                    pass
-            return res
-
 
     def run_command(self, command):
         self.current_code = command
@@ -972,51 +924,10 @@ due to the exec, so we do it here"""
             _debug_write('Unknown module ' + module)
 
     def get_module_names(self):
-        res = []
-        for name, module in sys.modules.items():
-            try:
-                if name != 'visualstudio_py_repl' and name != '$visualstudio_py_debugger':
-                    if sys.platform == 'cli' and type(module) is NamespaceType:
-                        self.get_namespaces(name, module, res)
-                    else:
-                        try:
-                            filename = os.path.abspath(module.__file__)
-                        except Exception:
-                            filename = None
-                        res.append((name, filename))
-
-            except:
-                pass
-        return res
-
-    def get_namespaces(self, basename, namespace, names):
-        names.append((basename, ''))
-        try:
-            for name in dir(namespace):
-                new_name = basename + '.' + name
-                new_namespace = getattr(namespace, name)
-
-                if type(new_namespace) is NamespaceType:
-                    self.get_namespaces(new_name, new_namespace, names)
-        except:
-            pass
+        return get_module_names()
 
     def flush(self):
         sys.stdout.flush()
-
-    def do_detach(self):
-        import visualstudio_py_debugger
-        visualstudio_py_debugger.DETACH_CALLBACKS.remove(self.do_detach)
-        self.on_debugger_detach()
-
-    def attach_process(self, port, debugger_id, debug_options):
-        def execute_attach_process_work_item():
-            import visualstudio_py_debugger
-            visualstudio_py_debugger.DETACH_CALLBACKS.append(self.do_detach)
-            visualstudio_py_debugger.attach_process(port, debugger_id, debug_options, report=True, block=True)
-
-        self.execute_item = execute_attach_process_work_item
-        self.execute_item_lock.release()
 
     @staticmethod
     def get_type_name(val):
@@ -1045,6 +956,73 @@ due to the exec, so we do it here"""
                 except:
                     pass
             return
+
+def get_cur_module_set():
+    """gets the set of modules avoiding exceptions if someone puts something
+    weird in there"""
+
+    try:
+        return set(sys.modules)
+    except:
+        res = set()
+        for name in sys.modules:
+            try:
+                res.add(name)
+            except:
+                pass
+        return res
+
+def get_module_names():
+    res = []
+    for name, module in sys.modules.items():
+        try:
+            if name != 'visualstudio_py_repl' and name != '$visualstudio_py_debugger':
+                if sys.platform == 'cli' and type(module) is NamespaceType:
+                    get_namespaces(name, module, res)
+                else:
+                    try:
+                        filename = os.path.abspath(module.__file__)
+                    except Exception:
+                        filename = None
+                    res.append((name, filename))
+
+        except:
+            pass
+    return res
+
+def get_namespaces(basename, namespace, names):
+    names.append((basename, ''))
+    try:
+        for name in dir(namespace):
+            new_name = basename + '.' + name
+            new_namespace = getattr(namespace, name)
+
+            if type(new_namespace) is NamespaceType:
+                get_namespaces(new_name, new_namespace, names)
+    except:
+        pass
+
+def print_exception_frames(exc_type, exc_value, exc_tb):
+    exc_next = skip_internal_frames(exc_tb)
+    sys.stderr.write(''.join(traceback.format_exception(exc_type, exc_value, exc_next)))
+
+def skip_internal_frames(tb):
+    """return the first frame outside of the repl/debugger code"""
+    while tb is not None and is_internal_frame(tb):
+        tb = tb.tb_next
+    return tb
+
+def is_internal_frame(tb):
+    """return true if the frame is from internal code (repl or debugger)"""
+    f = tb.tb_frame
+    co = f.f_code
+    filename = co.co_filename
+    return filename.endswith('visualstudio_py_repl.py') or filename.endswith('visualstudio_py_debugger.py')
+
+'''
+This code is no longer used.
+Kept as a reference for now until the members and signatures functionality is
+ported over to the debugger.
 
 class DebugReplBackend(BasicReplBackend):
     def __init__(self, debugger):
@@ -1177,6 +1155,7 @@ class DebugReplBackend(BasicReplBackend):
 
     def check_for_exit_execution_loop(self):
         return self.disconnect_requested
+'''
 
 class _ReplOutput(object):
     """file like object which redirects output to the repl window."""
