+++ Python/Tests/Core/DebugReplEvaluatorTests.cs
@@ -17,6 +17,7 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -36,6 +37,7 @@ namespace PythonToolsTests {
         private PythonDebugReplEvaluator _evaluator;
         private MockReplWindow _window;
         private List<PythonProcess> _processes;
+        private string _moduleFileExtension;
 
         [ClassInitialize]
         public static void DoDeployment(TestContext context) {
@@ -63,6 +65,7 @@ namespace PythonToolsTests {
             _window = new MockReplWindow(_evaluator);
             _evaluator._Initialize(_window);
             _processes = new List<PythonProcess>();
+            _moduleFileExtension = Version.Version < Microsoft.PythonTools.Parsing.PythonLanguageVersion.V30 && Version.IsCPython ? ".pyc" : ".py";
         }
 
         [TestCleanup]
@@ -158,6 +161,35 @@ NameError: name 'does_not_exist' is not defined
         }
 
         [TestMethod, Priority(1)]
+        public async Task AvailableScopes() {
+            await AttachAsync("DebugReplTest1.py", 3);
+
+            var expectedData = new Dictionary<string, string>() {
+                { "<Current Frame>", null },
+                { "abc", "abc" },
+                { "dis", "dis" },
+            };
+
+            Assert.IsTrue(_evaluator.EnableMultipleScopes);
+
+            var scopes = _evaluator.GetAvailableScopes().ToArray();
+            foreach (var expectedItem in expectedData) {
+                CollectionAssert.Contains(scopes, expectedItem.Key);
+            }
+
+            var scopesAndPaths = _evaluator.GetAvailableScopesAndPaths().ToArray();
+            foreach (var expectedItem in expectedData) {
+                var actualItem = scopesAndPaths.SingleOrDefault(d => d.Key == expectedItem.Key);
+                Assert.IsNotNull(actualItem);
+                if (!string.IsNullOrEmpty(expectedItem.Value)) {
+                    Assert.IsTrue(PathUtils.IsSamePath(Path.Combine(Version.PrefixPath, "lib", expectedItem.Value + _moduleFileExtension), actualItem.Value));
+                } else {
+                    Assert.IsNull(actualItem.Value);
+                }
+            }
+        }
+
+        [TestMethod, Priority(1)]
         public async Task ChangeModule() {
             await AttachAsync("DebugReplTest1.py", 3);
 
@@ -165,11 +197,23 @@ NameError: name 'does_not_exist' is not defined
 
             // Change to the dis module
             Assert.AreEqual("Current module changed to dis", ExecuteCommand(new SwitchModuleCommand(), "dis"));
+            Assert.AreEqual("dis", _evaluator.CurrentScopeName);
+            Assert.IsTrue(PathUtils.IsSamePath(_evaluator.CurrentScopePath, Path.Combine(Version.PrefixPath, "lib", "dis" + _moduleFileExtension)));
             Assert.AreEqual("", ExecuteText("test = 'world'"));
             Assert.AreEqual("'world'", ExecuteText("test"));
 
-            // Change back to the current frame
-            Assert.AreEqual("Current module changed to <CurrentFrame>", ExecuteCommand(new SwitchModuleCommand(), "<CurrentFrame>"));
+            // Change back to the current frame (using localized name)
+            Assert.AreEqual("Current module changed to <Current Frame>", ExecuteCommand(new SwitchModuleCommand(), "<Current Frame>"));
+            Assert.AreEqual("<Current Frame>", _evaluator.CurrentScopeName);
+            Assert.AreEqual("", _evaluator.CurrentScopePath);
+            Assert.AreEqual("", ExecuteText("test", false));
+            Assert.IsTrue(_window.Error.Contains("NameError:"));
+            Assert.AreEqual("'hello'", ExecuteText("a"));
+
+            // Change back to the current frame (using fixed and backwards compatible name)
+            Assert.AreEqual("Current module changed to <Current Frame>", ExecuteCommand(new SwitchModuleCommand(), "<CurrentFrame>"));
+            Assert.AreEqual("<Current Frame>", _evaluator.CurrentScopeName);
+            Assert.AreEqual("", _evaluator.CurrentScopePath);
             Assert.AreEqual("", ExecuteText("test", false));
             Assert.IsTrue(_window.Error.Contains("NameError:"));
             Assert.AreEqual("'hello'", ExecuteText("a"));
@@ -331,8 +375,7 @@ NameError: name 'does_not_exist' is not defined
             PythonProcess process = debugger.DebugProcess(Version, DebuggerTestPath + filename, async (newproc, newthread) => {
                 var breakPoint = newproc.AddBreakpointByFileExtension(lineNo, filename);
                 await breakPoint.AddAsync(TimeoutToken());
-            },
-            debugOptions: PythonDebugOptions.CreateNoWindow);
+            });
 
             _processes.Add(process);
