+++ Python/Tests/DebuggerTests/DebugExtensions.cs
@@ -33,7 +33,7 @@ namespace DebuggerTests {
             }
             var process = debugger.CreateProcess(version.Version, version.InterpreterPath, arguments, dir, "", interpreterOptions, debugOptions);
             process.DebuggerOutput += (sender, args) => {
-                Console.WriteLine("{0}: {1}", args.Thread.Id, args.Output);
+                Console.WriteLine("{0}: {1}", args.Thread?.Id, args.Output);
             };
             process.ProcessLoaded += async (sender, args) => {
                 if (onLoaded != null) {
